Rewrite vectors to use an RCU-style extra level of indirection, in preparation for a threadsafe reallocator to come.
Fix a few unininitialized data conditions with "dieArg", "error" and
"temps".

Put the "check GC lock" call in a special case JMPLOOP instruction to
avoid extra overhead.  This is guaranteed to be called exactly once
every loop iteration, which is the minimum overhead for a call that
needs to be made every few instructions (although I should check the
overhead of an instruction counter too...).