Do *not* check the bottleneck at the top of naModLock() before incrementing the thread count.  Until the thread count is incremented, we are a phantom thread that the GC does not know about!  Doing this here will cause us to "steal" the sleep position of another thread and the GC will start one thread too early.
Remove some ancient debugging code (freed object check) from gc.c.

A few spots have been decomposed into sub-functions to provide the
optimizer less opportunity to save must-to-be-stored-for-threadsafety
values in registers.  It's not at all clear to me that this was
necessary, but the resulting code is somewhat cleaner, so these
changes can stay.

Fixed a few correctness isssues with hash/vector reallocations.  If
you ever get a 3-way collision (3+ threads trying to simultaneously
insert into a full container) you need to loop inside the
reallocation.  A single if() is not sufficient.  Found by inspection.

Replace the temps vector with a hand-maintained array of pointers.
There was a race where a reallocation of that vector could invoke the
bottle neck and cause GC that would clean up the "new" version of the
array before it got swapped into the vector.  This is actually a much
better implementation anyway, as it avoids thread contention on the
(never shared between threads by definition, and *very* commonly used)
temps vector.