Continue reporting "no error" as a NULL return from naGetError().
Make the very top level function in naCall() set its namespace to the
local variables.  This isn't completely symmetric, normally such a
function's namespace would point to the local variables of the scope
in which it was assigned (which doesn't exist here).  But this allows
code to use caller()[1] as the "next" argument to bind and have it
work correctly at the top level.