Banged out a simple (but surprisingly complete) apache module.  No docs yet, nor much in the way of test code.  But it seems to work.
Check in some work on the template.nas file, which was the impetus for
the Apache module.

Add a user data pointer to the context, for call-specific data that I
don't want to wrap with a ghost (e.g. an Apache request_rec).

Fix warning condition with code.c:getMember()

Blank the dieArg when returning from a function call.  It's useless at
that point, and this allows it to be collected.

We create the stdin/out/err handles in the io library as conveniences,
but we obviously don't own those FILE pointers.  If these ghosts get
collected, don't close them!

Fix ungetc() warning condition in iolib.c.

Yank the huge endianness detection logic in favor of a simple list of
platforms.  I'm sure I broke someone here, but I'm confident this will
be easier to maintain over the long term.

Clean up comments on the modlock, which implied that read-only access
without the lock was always safe (it's not -- the references can be
collected out from under you if another thread drops them).