Fix bad function name in interactive.nas.
Don't do a tail call from the top frame, otherwise caller() can be in
the odd position of returning nil to a function that clearly got
called from somewhere.

Better error checking in the code generator to fix a crash when it
tried to generate code for bad syntax (e.g. a TOK_MINUS with no
children).

Docs for the preliminary readline library.  This might get replaced by
Jonatan in the future, but might as well do it so it doesn't get
forgotten.