Fix argument ordering issues in regex sub/sub_all functions. Change naming of group variables in substitute expressions.
Add a soft-coded readfile() function to the io module.

Allow ghost destroy callbacks to be null (and ignored) instead of
causing a crash.

Include the readline library under the right HAVE_* symbol.

Add a workaround to what appears (?) to be a PCRE bug with the
startoffset parameter to pcre_exec.  Do the offset processing
outside the call.

Update the apache module to use the module system via driver.nas.
The gethdr() and getcgi() builtins now return a vector of all
keys when called with no arguments.  Update the sample code and
README to be a little more illustrative.  Change the ".nas"
convention for "servlet" files to ".ns" to avoid colliding with
regular nasal source files on the same server (i.e. sample.nas).