In nasal.h, add a "pure" attribute when compiling under GCC.  That allows the compiler to better optimize pervasive calls to the naIs*() predicates, etc...
Add GtkDialog facilities (the vbox and action_area members, and a
dialog_run() method) and a "msgdialog.nas" example.

Misc cleanup of the gtk code to unify it with the rest of Nasal, and
in some cases to shrink it considerably (net -30 lines of code).
Hopefully I didn't break anything.