Finaly nifty feature for now: vector slicing.  You can now generate a vector from a list of indices or ranges, e.g.:
  var (uid, gid, size) = io.stat(filename)[4,5,7];
  var tail = list[1:];
  var prefix = list[0:length];

The two values of a range are the start and end indices, inclusive.
They can be runtime expressions -- they don't need to be constants.
As for vector indexing otherwise, negative indices are relative to the
length of the list (i.e. -1 refers to the last element).  Missing
sides to the range are interpreted as either 0 (":N") or -1 ("N:").